+++ massiv/src/Data/Massiv/Array/Stencil.hs
@@ -91,4 +91,3 @@ makeStencil !sSz !sCenter relStencil =
 {-# INLINE makeStencil #-}
 
 
-